fix(toolbar-editor): rename component files
- editor-toolbar -> toolbar-editor
- ReactiveNavigation -> ToolbarEditor
