+++ packages/sage-assets/lib/stylesheets/core/mixins/_sage.scss
@@ -136,7 +136,7 @@
         }
 
         &.sage-btn--rich-text,
-        .editor-toolbar__section & {
+        .toolbar-editor__section & {
           margin-left: sage-spacing(3xs);
           margin-right: sage-spacing(3xs);
           padding: sage-spacing(xs) rem(12px) sage-spacing(xs) sage-spacing(xs);
